refactor create image
