+++ vite-project/src/Modules/CreateImage/components/RandomPrompt.tsx
@@ -1,5 +1,5 @@
-import Button from "../Button";
-import Card from "../Card";
+import Button from "../../../components/Button";
+import Card from "../../../components/Card";
 
 interface RandomPromptProps {
   handleRandomPrompt: () => void;
