+++ vite-project/src/lib/api.ts
@@ -1,5 +1,3 @@
-import { postData } from "../Modules/DisplayImages/lib/types";
-
 export const fetchUser = async (email: string, token: string) => {
   const res = await fetch(`${import.meta.env.VITE_API_URL}/api/v1/user/`, {
     method: "GET",
@@ -34,73 +32,6 @@ export const favouriteImage = async (id: string, token: string) => {
   } else {
     throw new Error("Something went wrong");
   }
-};
-
-export const createPost = async (data: postData) => {
-  try {
-    const res = await fetch(`${import.meta.env.VITE_API_URL}/api/v1/post`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    });
-
-    if (res.ok) {
-      const result = await res.json();
-      return result;
-    } else {
-      throw new Error("Something went wrong with create post request");
-    }
-  } catch (e) {
-    console.log(e, "create post error");
-  }
-};
-
-export const initalGptPrompt = () => {
-  return fetch(`${import.meta.env.VITE_API_URL}/api/v1/gpt/initial`, {
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  });
-};
-
-export const uploadAudioFile = async (file: File | Blob) => {
-  const formData = new FormData();
-  formData.append("file", file);
-
-  const res = await fetch(
-    `${import.meta.env.VITE_API_URL}/api/v1/whisper/upload`,
-    {
-      method: "POST",
-      body: formData,
-    }
-  );
-
-  return res.json();
-};
-
-export const getWhisperTranscript = async () => {
-  const res = await fetch(`${import.meta.env.VITE_API_URL}/api/v1/whisper/`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-  });
-  const data = await res.json();
-  console.log(data);
-  return data;
-};
-
-export const getGptPrompt = async (chips: string[]) => {
-  const prompt = await fetch(`${import.meta.env.VITE_API_URL}/api/v1/gpt`, {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ chips }),
-  });
-
-  return prompt.json();
 };
 
 export const sendTwilioText = async (
