+++ vite-project/src/pages/index.ts
@@ -1,4 +1,4 @@
 import Home from "../Modules/DisplayImages/pages/Home";
-import CreatePost from "./CreatePost";
+import CreatePost from "../Modules/CreateImage/pages/CreatePost";
 
 export { Home, CreatePost };
